requirements.txt
discord.py
asyncpraw
pillow
pandas
opencv-python-headless
async_google_trans_new
discord-ext-paginator
bs4